% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingMILEM.R
\name{testingMILEM}
\alias{testingMILEM}
\title{testingMILEM}
\usage{
testingMILEM(MILdatTst, trainRes)
}
\arguments{
\item{trainRes}{Results results from trainging results}

\item{testData}{testing data set}
}
\value{
YHat the empirical risk of the test
}
\description{
Compute the empirical risk of cv or testing for kernel machies with missing covariates
}
