% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingMILEM_randomK.R
\name{testingMILEM_randomK}
\alias{testingMILEM_randomK}
\title{testingMILEM_randomK}
\usage{
testingMILEM_randomK(MILdatTst, trainRes)
}
\arguments{
\item{MILdatTst}{testing data set}

\item{trainRes}{Results results from trainging results}
}
\value{
YHat the empirical risk of the test
}
\description{
Compute the empirical risk of cv or testing for MIL with multiple instance
}
